+++ lazyleech/plugins/leech.py
@@ -17,8 +17,7 @@ async def torrent_cmd(client, message):
     send_as_zip = 'zip' in text.pop(0)
     link = None
     reply = message.reply_to_message
-    document = message.document
-    if document:
+    if document := message.document:
         if document.file_size < 1048576 and document.file_name.endswith('.torrent') and (not document.mime_type or document.mime_type == 'application/x-bittorrent'):
             os.makedirs(str(message.from_user.id), exist_ok=True)
             fd, link = tempfile.mkstemp(dir=str(message.from_user.id), suffix='.torrent')
@@ -34,16 +33,23 @@ async def torrent_cmd(client, message):
         elif not getattr(reply, 'empty', True):
             document = reply.document
             link = reply.text
-            if document:
-                if document.file_size < 1048576 and document.file_name.endswith('.torrent') and (not document.mime_type or document.mime_type == 'application/x-bittorrent'):
-                    os.makedirs(str(message.from_user.id), exist_ok=True)
-                    fd, link = tempfile.mkstemp(dir=str(message.from_user.id), suffix='.torrent')
-                    os.fdopen(fd).close()
-                    await reply.download(link)
-                    mimetype = await get_file_mimetype(link)
-                    if mimetype != 'application/x-bittorrent':
-                        os.remove(link)
-                        link = reply.text or reply.caption
+            if (
+                document
+                and document.file_size < 1048576
+                and document.file_name.endswith('.torrent')
+                and (
+                    not document.mime_type
+                    or document.mime_type == 'application/x-bittorrent'
+                )
+            ):
+                os.makedirs(str(message.from_user.id), exist_ok=True)
+                fd, link = tempfile.mkstemp(dir=str(message.from_user.id), suffix='.torrent')
+                os.fdopen(fd).close()
+                await reply.download(link)
+                mimetype = await get_file_mimetype(link)
+                if mimetype != 'application/x-bittorrent':
+                    os.remove(link)
+                    link = reply.text or reply.caption
     if not link:
         await message.reply_text('''Usage:
 - /torrent <i>&lt;Torrent URL or File&gt;</i>
@@ -150,7 +156,7 @@ async def initiate_directdl(client, message, link, filename, send_as_zip):
     else:
         await handle_leech(client, message, gid, reply, user_id, send_as_zip)
 
-leech_statuses = dict()
+leech_statuses = {}
 async def handle_leech(client, message, gid, reply, user_id, send_as_zip):
     prevtext = None
     torrent_info = await aria2_tell_status(session, gid)
@@ -221,7 +227,11 @@ async def handle_leech(client, message, gid, reply, user_id, send_as_zip):
         try:
             await aria2_remove(session, gid)
         except Aria2Error as ex:
-            if not (ex.error_code == 1 and ex.error_message == f'Active Download not found for GID#{gid}'):
+            if (
+                ex.error_code != 1
+                or ex.error_message
+                != f'Active Download not found for GID#{gid}'
+            ):
                 raise
         finally:
             if task:
@@ -267,24 +277,21 @@ async def cancel_leech(client, message):
         gid = text[0].strip()
     elif not getattr(reply, 'empty', True):
         reply_identifier = (reply.chat.id, reply.message_id)
-        task = upload_statuses.get(reply_identifier)
-        if task:
+        if task := upload_statuses.get(reply_identifier):
             task, starter_id = task
             if message.chat.id not in ADMIN_CHATS and user_id != starter_id:
                 await message.reply_text('You did not start this leech.')
             else:
                 task.cancel()
             return
-        result = progress_callback_data.get(reply_identifier)
-        if result:
+        if result := progress_callback_data.get(reply_identifier):
             if message.chat.id not in ADMIN_CHATS and user_id != result[3]:
                 await message.reply_text('You did not start this leech.')
             else:
                 stop_uploads.add(reply_identifier)
                 await message.reply_text('Leech Cancelled!🤧')
             return
-        starter_id = upload_waits.get(reply_identifier)
-        if starter_id:
+        if starter_id := upload_waits.get(reply_identifier):
             if message.chat.id not in ADMIN_CHATS and user_id != starter_id[0]:
                 await message.reply_text('You did not start this leech.')
             else:
